2009-09-16 18:10 UTC+0100 Miguel Angel Marchuet <miguelangel@marchuet.net>    * compile.mak    * utils\hbfilere\hbfilere.c    * utils\hbfilere\hbfilere.h      + Added TransmitPackets that increase transfer speed. Remember that is needed to build        with versions of borland 5.8 or newer, or msvc compiler.    * source\common\hbver.c      + Added recognision of more versions of borland compilers.    * source\rtl\hbip.c      - removed unneeded code.    * source\rtl\filenet.c    * source\rtl\session.c    * source\rdd\dbcmd.c      ! fixed some warnings.

+++ xharbour/source/rtl/session.c
@@ -1,5 +1,5 @@
 /*
- * $Id: session.c,v 1.2 2009/07/22 16:55:13 marchuet Exp $
+ * $Id: session.c,v 1.3 2009/07/29 17:15:54 marchuet Exp $
  */
 
 /*
@@ -113,7 +113,7 @@ int hb_sessionID( void )
 {
    HB_TRACE( HB_TR_DEBUG, ( "hb_sessionID()" ) );
 
-   return pSession ? ( int ) pSession->id : NULL;
+   return pSession ? ( int ) pSession->id : ( int ) NULL;
 }
 
 int hb_sessionSelectID( const int sessionID )
